Set up init insert all end point POST, also changed point to be array of double

+++ libs/kdtree/include/node.hpp
@@ -13,7 +13,7 @@
 namespace sortedkdtree{
 
     //2D Tree
-    using Point =  std::array<float,2>;
+    using Point =  std::array<double,2>;
     using PointVec = std::vector<Point>;
